+++ components/nimbus/tests/common/mod.rs
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 #![cfg(feature = "rkv-safe-mode")]
 
-use remote_settings::{RemoteSettingsConfig2, RemoteSettingsContext, RemoteSettingsService};
+use remote_settings::{RemoteSettingsConfig, RemoteSettingsContext, RemoteSettingsService};
 use rkv::StoreOptions;
 
 // utilities shared between tests
@@ -84,7 +84,7 @@ fn new_test_client_internal(
         ..Default::default()
     };
 
-    let config = RemoteSettingsConfig2 {
+    let config = RemoteSettingsConfig {
         server: Some(RemoteSettingsServer::Custom {
             url: url.as_str().to_string(),
         }),